multimedia/banshee: Update to 2.2.1.
I though we had this in the FreeBSD tree but I realise only now that this
version is only in my svn checkout and not committed.  2.4 is out for a while,
so I'll try to update this port to the latest version.  I push this one since
with mono-2.11.1 (in the FreeBSD ports) the freezes I always got seems to be
gone (finger crossed!)